==> 5.Behavioral_classification.Rmd <==
---
title: "HMM_behaviors"
author: "Camila Calderon"
date: "2024-07-22"
output: html_document
---

## Load libraries

```{r setup}
library(momentuHMM)
library(tidyverse)
library(lubridate)
library(parallel)
library(mixtools)
library(plotly)
library(ggmap)
library(adehabitatLT)
```

## HMM for all data where GPS fixes GPS sampling is every 2 minutes 

```{r, results=FALSE}
# load data
load(file="~/ownCloud/PhDLife/P.hastatus/Thesis/paper1/BiologyLetters/data/allPhastBocas_merged.RData")

# if step and angle are present need to be removed
names(allbats_Bocas)
allbats_Bocas <- allbats_Bocas[!names(allbats_Bocas) %in% c("step", "angle")]

# order by timestamp and Id
allbats_Bocas <- arrange(allbats_Bocas, ID_batday,timestamp)

# create ADE trajectory added with NAs every 2 minutes for all individuals expect for individuals in March. I decided not to interpolate as it was creating unrealistic tracks. 

# filter
all_but_march <- allbats_Bocas %>%
  filter(date<="2022-02-27" | date>="2022-03-30")

# add NA's every 2 minutes
df_na <- setNA(ltraj = as.ltraj(xy=all_but_march[, c("x", "y")], date = all_but_march$timestamp, id=all_but_march$ID_batday), date.ref =all_but_march$timestamp[1], dt=2, tol=2, units="min")

# convert ltraj object to dataframe
df <- ld(df_na)[,c("x","y", "id","date")]

# split data frame by idday
bats.split <- split(df, df$id)

# process data with Prepdata
batsData<- lapply(bats.split, function(x){
  momentuHMM::prepData(x, type = 'UTM', coordNames = c('x', 'y'))
})

# rename the id column with right id
batsData <- lapply(1:length(batsData), function(i){
  id_day <- names(batsData[i])
  batsData[[i]]$ID <- id_day
  return(batsData[[i]])
})

# to check steps one id
hist(batsData[[40]]$step)

# plotting to see step lengths and turning angles
pdf(file="~/ownCloud/PhDLife/P.hastatus/Thesis/Paper1/BiologyLetters/supplementary/HMM_outcomes/Parameters_nointerpol_nomarch.pdf")
lapply(batsData, function(x){
  tryCatch({
    plot(x, ask=FALSE)
  },error=function(e) finally = print("f'd up"))
})
dev.off()
```

## Classify behaviors: foraging and commuting with a HMM 2 state model

```{r pressure, results=FALSE}
# estimate parameters for HMM
# selecting step lengths and turning angles for states model
batsData.df <- do.call(rbind, batsData)

# calculates normal distribution from data and select 2 values from it
mixHMM2 <- normalmixEM(na.omit(batsData.df$step), k = 2)
# plot(mixHMM,which=3)
idx2 <- order(mixHMM2$mu)
# select step length values
mu02 <- c(mixHMM2$mu[idx2[1]], mixHMM2$mu[idx2[2]])
# select step SD
sigma02 <- c(mixHMM2$sigma[idx2[1]],  mixHMM2$sigma[idx2[2]]) # step SD


## Fitting HMM with 2 states
state2Names <- c("state1", "state2")
# initial parameters
step2ParMean <- mu02
step2ParSD <- sigma02
zero2mass0 <- c(0.01, 0.01)
step2Par0 <- c(step2ParMean,step2ParSD, zero2mass0)
angle2Par0 <- c(0.5,0.1)
dist = list(step = "gamma", angle = "wrpcauchy")
m <- fitHMM(data=batsData.df, 
            nbStates=2, 
            dist=dist,
            Par0=list(step=step2Par0, angle=angle2Par0),
            stateNames = state2Names)

# save results 
pdf(file="~/ownCloud/PhDLife/P.hastatus/Thesis/Paper1/BiologyLetters/supplementary/HMM_outcomes/HMM_2states_Bocas_nomarch.pdf")
plot(m, plotCI = TRUE, ask=FALSE)
dev.off()
```

## Checking classification from the model and add it with the original data frame
```{r, results=FALSE}
# viterbi function reconstructs most probably states sequence for each bat
m$data$HMM2 <- viterbi(m)
m$data$HMM2_prob <- stateProbs(m)

# order data again
all_but_march <- all_but_march[order(all_but_march$ID_batday, all_but_march$timestamp),]

#remove the NAs that were included in the beggining
m_df <- m$data[which(!is.na(m$data$x)),]

# adding the states to the initial  dataframe
all_but_march$HMM <- m_df$HMM2[match(all_but_march$timestamp, m_df$date)]
all_but_march$HMMprob <- m_df$HMM2HMM2_prob[match(all_but_march$timestamp, m_df$date)]
all_but_march$turningangle <- m_df$angle[match(all_but_march$timestamp, m_df$date)]
all_but_march$step <- m_df$step[match(all_but_march$timestamp, m_df$date)]
```

## Change classification points which are wrongly assigned in either of the behaviors

```{r, results=FALSE}
# change state numbers to behaviors and adjust manually some of the classification
all_but_march$behav <- all_but_march$HMM
all_but_march$behav[which(all_but_march$HMM==1)] <- "foraging"
all_but_march$behav[which(all_but_march$HMM==2)] <- "commuting"
unique(all_but_march$behav)
all_but_march[which(is.na(all_but_march$behav)),]
all_but_march$behav[(all_but_march$behav=="commuting" & all_but_march$ground_speed<1)] <- "foraging"
# some points that are flying are classified as foraging
all_but_march$behav[(all_but_march$behav=="foraging" & all_but_march$ground_speed>5)] <- "commuting"
#these points are wrongly classified
all_but_march$behav[(all_but_march$behav=="foraging" & all_but_march$ID== "2D507235_G_2022-02-04" & all_but_march$x>=350000)] <- "commuting"

# checking for na's
all_but_march <- all_but_march[which(!is.na(all_but_march$behav)),]

# plot foraging points results
mapview::mapView(all_but_march[all_but_march$behav=="foraging",], xcol="location_long", ycol="location_lat", zcol="behav", legend=F, crs="EPSG:4326") #as points

# event id's which are in the water seem more like commuting 
all_but_march$behav[all_but_march$event_id=="29640571357" | all_but_march$event_id=="29640585899" | all_but_march$event_id=="29640555868"| all_but_march$event_id=="29640511978" | all_but_march$event_id=="29640589130"| all_but_march$event_id=="29640513640" | all_but_march$event_id=="29640503867" | all_but_march$event_id=="1184403553" ] <- "commuting"

					
# plot foraging points results
mapview::mapView(all_but_march[all_but_march$behav=="foraging",], xcol="location_long", ycol="location_lat", zcol="behav", legend=F, crs="EPSG:4326") #as points

# look at the histograms of speed and turning angle
hist(all_but_march[all_but_march$behav=="commuting",]$ground_speed)
hist(all_but_march[all_but_march$behav=="foraging",]$ground_speed)
hist(all_but_march[all_but_march$behav=="commuting",]$turningangle)
hist(all_but_march[all_but_march$behav=="foraging",]$turningangle)
```

## HMM for the March 2022 data where GPS fixes were taken every 3 minutes

```{r}
# filter only march data
march <- allbats_Bocas %>%
  filter(date>="2022-02-27" & date<="2022-03-30")

# add na's to regularize sampling
df_na_march <- setNA(ltraj = as.ltraj(xy=march[, c("x", "y")], date = march$timestamp, id=march$ID_batday), date.ref =march$timestamp[1], dt=3, tol=3, units="min")

# convert ltraj object to dataframe
df_march <- ld(df_na_march)[,c("x","y", "id","date")]

# split data frame by idday
bats_split_march <- split(df_march, df_march$id)

# process data with Prepdata
batsData_march<- lapply(bats_split_march, function(x){
  momentuHMM::prepData(x, type = 'UTM', coordNames = c('x', 'y'))
})

# rename the id column with right id
batsData_march <- lapply(1:length(batsData_march), function(i){
  id_day <- names(batsData_march[i])
  batsData_march[[i]]$ID <- id_day
  return(batsData_march[[i]])
})

# to check steps one id
hist(batsData_march[[100]]$step)

# plotting to see step lengths and turning angles
pdf(file="~/ownCloud/PhDLife/P.hastatus/Thesis/Paper1/BiologyLetters/supplementary/HMM_outcomes/Parameters_nointerpol_march.pdf")
lapply(batsData_march, function(x){
  tryCatch({
    plot(x, ask=FALSE)
  },error=function(e) finally = print("f'd up"))
})
dev.off()

# estimate parameters for HMM
# selecting step lengths and turning angles for states model
batsDatamarch_df <- do.call(rbind, batsData_march)

# calculates normal distribution from data and select 2 values from it
mixHMM2 <- normalmixEM(na.omit(batsDatamarch_df$step), k = 2)
# plot(mixHMM,which=3)
idx2 <- order(mixHMM2$mu)
# select step length values
mu02 <- c(mixHMM2$mu[idx2[1]], mixHMM2$mu[idx2[2]])
# select step SD
sigma02 <- c(mixHMM2$sigma[idx2[1]],  mixHMM2$sigma[idx2[2]]) # step SD


# fitting HMM for whole individuals with 2 states
state2Names <- c("state1", "state2")
# initial parameters
step2ParMean <- mu02
step2ParSD <- sigma02
zero2mass0 <- c(0.01, 0.01)
step2Par0 <- c(step2ParMean,step2ParSD, zero2mass0)
angle2Par0 <- c(0.5,0.1)
dist = list(step = "gamma", angle = "wrpcauchy")
m2_march <- fitHMM(data=batsDatamarch_df, 
                   nbStates=2, 
                   dist=dist,
                   Par0=list(step=step2Par0, angle=angle2Par0),
                   stateNames = state2Names)
m2_march

pdf(file="~/ownCloud/PhDLife/P.hastatus/Thesis/Paper1/BiologyLetters/supplementary/HMM_outcomes/HMM_2states_Bocas_march.pdf")
plot(m2_march, plotCI = TRUE, ask=FALSE)
dev.off()
```
## Assign behavior to the states

```{r}
# viterbi function reconstructs most probably states sequence for each bat
m2_march$data$HMM2 <- viterbi(m2_march)
m2_march$data$HMM2_prob <- stateProbs(m2_march)

# order data again
march <- march[order(march$ID_batday, march$timestamp),]

# remove the NAs that were included in the beggining
m2_df_march <- m2_march$data[which(!is.na(m2_march$data$x)),]

# adding the states to the march dataframe
march$HMM <- m2_df_march$HMM2[match(march$timestamp, m2_df_march$date)]
march$HMMprob <- m2_df_march$HMM2HMM2_prob[match(march$timestamp, m2_df_march$date)]
march$turningangle <- m2_df_march$angle[match(march$timestamp, m2_df_march$date)]
march$step <- m2_df_march$step[match(march$timestamp, m2_df_march$date)]

# model fit can be assessed using the pseudo-residuals, with the functions pseudoRes and plotPR
r_march <- pseudoRes(m2_march)
plotPR(m2_march)
acf(r_march$stepRes[!is.na(r_march$stepRes)], lag.max = 60)

# change state numbers to behaviors and adjust manually some of the classification
march$behav <- march$HMM
march$behav[which(march$HMM==2)] <- "commuting"
unique(march$behav)

march$behav[which(march$HMM==1)] <- "foraging"
march$behav[(march$behav=="commuting" & march$ground_speed<1)] <- "foraging"
march$behav[(march$behav=="foraging" & march$ground_speed>5)] <- "commuting"

#remove na's
march <- march[which(!is.na(march$behav)),]

# plot first results
mapview::mapView(march[march$behav=="foraging",], xcol="location_long", ycol="location_lat", zcol="tag_local_identifier", legend=F, crs="EPSG:4326") #as points

#change individual point 
march$behav[march$event_id=="22068505980"] <- "commuting"

# look at the histograms
hist(march[march$behav=="commuting",]$ground_speed)
hist(march[march$behav=="foraging",]$ground_speed)
hist(march[march$behav=="commuting",]$turningangle)
hist(march[march$behav=="foraging",]$turningangle)

# merge both data frames (march and all the rest of the data) after HMM
bats_behaviors <- rbind(all_but_march, march)

# save
save(bats_behaviors,file="~/ownCloud/PhDLife/P.hastatus/Thesis/Paper1/BiologyLetters/data/Phyllostomus_HMMbehaviors.RData")
```
