==> 7.Straightness.Rmd <==
---
Author: "Camila Calderon"
title: "Turning angles"
date: "2023-07-22"
---

# Straightness of commuting tracks for all bats in Bocas

## Loading packages

```{r setup, global_options, include=FALSE}
library(lubridate)
library(dplyr)
library(ggplot2)
library(move)
library(circular)
library(patchwork) 
library(units)
```

## Loading data

```{r loading, include=FALSE}
load(file="~/ownCloud/PhDLife/P.hastatus/Thesis/Paper1/BiologyLetters/data/Phyllostomus_HMMbehaviors.RData")

head(bats_behaviors) 
unique(bats_behaviors$ID)
bats_behaviors  <- bats_behaviors[order(bats_behaviors$tag_local_identifier, bats_behaviors$timestamp),]
```

## Add speed and turning angles to the data frame

```{r add speed and turning angle}
phasmove <-  move::move(x=bats_behaviors$location_long, y=bats_behaviors$location_lat,
     time=as.POSIXct(bats_behaviors$timestamp, format="%Y-%m-%d %H:%M:%S", tz="UTC"), 
     data=bats_behaviors, proj=CRS("+proj=longlat +ellps=WGS84"), 
     animal=bats_behaviors$tag_local_identifier, sensor="GPS")
n.indiv(phasmove)

# re do turning angles and speed after behavioral classification
moveList <- lapply(split(phasmove), function(myInd){
  myInd$tlag <- c(NA,timeLag(myInd, units="secs"))
  myInd$step <- c(NA, move::distance(myInd))
  myInd$tag_ground_speed <- myInd$ground_speed
  myInd$ground_speed <- c(NA, move::speed(myInd))
  myInd$angle <- c(NA, move::angle(myInd))
  return(myInd)
})
phas <- moveStack(moveList, forceTz="UTC")
phas <- spTransform(phas, CRS("+proj=utm +zone=17 +datum=WGS84"))
allbats_Bocas <- as.data.frame(phas)
```

## Select commuting tracks from the colony groups to the foraging spots

```{r select commutes, message=FALSE, warning=FALSE, results=FALSE}
# adding a variable that counts the length of consecutive commuting and foraging points. This will be used to filter only commuting tracks that contain more than 5 points in a row
allbats_Bocas$n <- NA
r <- rle(allbats_Bocas$behav)
i=1
for(i in 1:length(r$lengths)){
  #if(r$values[i] == "commuting"){
      start <- {}
      end <- {}
      try(start <- sum(r$lengths[1:(i-1)])+1)
      if(i == 1) start <- 1
      end <- sum(r$lengths[1:i])

      allbats_Bocas$n[start:end] <- r$lengths[i]
}

# filter only commuting with less than 5 points in a row
bats_commu <- allbats_Bocas %>% 
  filter(behav=="commuting" & n>=5)

# add time difference to split commuting tracks by the ID_day
bats_commu <- bats_commu %>%
  group_by(ID_batday) %>%
  mutate(diff = timestamps - lag(timestamps))
bats_commu$diff <- as.numeric(bats_commu$diff)*60
bats_commu$diff[is.na(bats_commu$diff)] <- 0

# data frame with data from 2016-February 2022
Commu_feb <- bats_commu %>% 
 filter(date(timestamp) < "2022-03-01")

# removing individuals with only one commuting track instead of two
Commu_feb <- Commu_feb %>% 
  filter(ID_batday!="0C506E35_C_3" , ID_batday!="74D972D_1" , ID_batday!= "71A0D95_2" , ID_batday!="2016030705_1" , ID_batday!="2016030703_2")
unique(Commu_feb$ID_batday)

# splitting tracks by ID
commu.split <- split(Commu_feb, f=Commu_feb$ID)

# for loop to divide each day of the animal in commuting outbound and commuting inbound using 30 minutes and 66 minutes as a threshold value
mylist <- list() 
for (i in 1:length(commu.split)){
  x <- which(commu.split[[i]]$diff>=1750 & commu.split[[i]]$diff<=2000 | commu.split[[i]]$diff>=4000)#70 minutes in seconds
  mylist[[i]] <- split(commu.split[[i]], cumsum(1:nrow(commu.split[[i]])%in%x))
  indv <- unique(names(commu.split[i]))
  names(mylist[[i]]) <- paste(indv, seq_along(mylist[[i]]), sep = "_")
  }

# convert nested list to data frame
find_df <- function(x) {
  if (is.data.frame(x))
    return(list(x))
  if (!is.list(x))
    return(NULL)
  unlist(lapply(x, find_df), FALSE)
}

data <- find_df(mylist)

# add names to list
f <- function (data, name){
  data$name <- name
  data
}

data <- Map(f, data, names(data)) 

# convert list to dataframe
data.commu <- as.data.frame(do.call(rbind, data))

# select only outbound commuting
data.commu.out <- data.commu[grep("_1", data.commu$name),]
unique(data.commu.out$name)  

# plot outbound commutes
ggplot(aes(x=location_long, y=location_lat), data=data.commu.out)+geom_point()

# filter to data that will be used
data.commu.out <- data.commu.out[!c(data.commu.out$date=="2021-12-12" | data.commu.out$date=="2021-12-11"),]
unique(data.commu.out$date)
```

## Select commutes in females and males in March 2022

```{r select commuting locations, results=FALSE}
Commu_march <- bats_commu %>% 
 filter(date> "2022-03-01", date< "2022-03-20")

# split 
splitmarch <- split(Commu_march, f=Commu_march$ID)

#complete tracks
Commu_march_complete <- splitmarch[names(splitmarch) %in% c("PH_TS_018_2022-03-08",  "PH_TS_018_2022-03-10", "PH_TS_039_2022-03-09", "PH_TS_046_2022-03-08", "PH_TS_046_2022-03-09", "PH_TS_056_2022-03-08", "PH_TS_072_2022-03-12", "PH_TS_072_2022-03-19", "PH_TS_072_2022-03-20", "PH_TS_081_2022-03-08", "PH_TS_081_2022-03-09", "PH_TS_096_2022-03-11", "PH_TS_103_2022-03-17", "PH_TS_011_2022-03-08", "PH_TS_011_2022-03-12", "PH_TS_011_2022-03-15", "PH_TS_011_2022-03-17", "PH_TS_014_2022-03-07", "PH_TS_014_2022-03-08",  "PH_TS_014_2022-03-09", "PH_TS_014_2022-03-10", "PH_TS_016_2022-03-09", "PH_TS_016_2022-03-10", "PH_TS_074_2022-03-08", "PH_TS_074_2022-03-10", "PH_TS_074_2022-03-12", "PH_TS_079_2022-03-07", "PH_TS_079_2022-03-16", "PH_TS_080_2022-03-09", "PH_TS_080_2022-03-10", "PH_TS_080_2022-03-11","PH_TS_080_2022-03-14", "PH_TS_080_2022-03-15" , "PH_TS_080_2022-03-16","PH_TS_080_2022-03-17", "PH_TS_080_2022-03-18", "PH_TS_080_2022-03-19", "PH_TS_080_2022-03-20", "PH_TS_080_2022-03-21", "PH_TS_100_2022-03-07", "PH_TS_100_2022-03-10", "PH_TS_100_2022-03-11", "PH_TS_100_2022-03-16", "PH_TS_100_2022-03-17", "PH_TS_100_2022-03-18", "PH_TS_112_2022-03-08", "PH_TS_112_2022-03-11", "PH_TS_113_2022-03-09", "PH_TS_120_2022-03-08", "PH_TS_120_2022-03-11") == TRUE] 

# for loop to divide each day of the animal in commuting outbound and commuting inbound using 70 minutes as a threshold value
mylist2 <- list() 
for (i in 1:length(Commu_march_complete)){
  x <- which(Commu_march_complete[[i]]$diff>=4200)
  mylist2[[i]] <- split(Commu_march_complete[[i]], cumsum(1:nrow(Commu_march_complete[[i]])%in%x))
  indv <- unique(names(Commu_march_complete[i]))
  names(mylist2[[i]]) <- paste(indv, seq_along(mylist2[[i]]), sep = "_")
  }

# convert nested list to data frame
data_march <- find_df(mylist2)

# add names to list of the commuting outbound or inbound to the list
data_march <- Map(f, data_march, names(data_march))

#replace PH_TS_072_2022-03-12_1 with PH_TS_072_2022-03-12_2 because _2 is the outbound track
data_march[["PH_TS_072_2022-03-12_1"]] <- data_march[["PH_TS_072_2022-03-12_2"]] 
# convert list to dataframe
data.commu.march <- as.data.frame(do.call(rbind, data_march))

# select only outbound commuting
data.commu.march.out <- data.commu.march[grep("_1\\b", data.commu.march$name),]#exact match to avoid selecting the inbound commuting tracks
data.commu.march.in <- data.commu.march[grep("_2\\b", data.commu.march$name),]

# plot lat and long
ggplot(aes(x=location_long, y=location_lat, color=ID_batday), data=data.commu.march.out)+geom_point()#they are OK

```

## Select commuting tracks for individuals in the wet season in 2023

```{r turning angle Bocas 2023, results=FALSE}
# filter data
commu_wet <- bats_commu %>%
  filter(date>="2023-08-01")

# checking commuting of this bat
ggplot(commu_wet, aes(x=location_long, y=location_lat, col=behav))+geom_point()

# split by ID
splitwet <- split(commu_wet, f=commu_wet$ID)

#complete tracks
commu_wet_complete <- splitwet[names(splitwet) %in% c("PHYL1_2023-08-13", "PHYL1_2023-08-14", "PHYL1_2023-08-15", "PHYL1_2023-08-17", "PHYL11_2023-08-15", "PHYL11_2023-08-16", "PHYL11_2023-08-17", "PHYL11_2023-08-18", "PHYL16_2023-08-14", "PHYL16_2023-08-15",  "PHYL21_2023-08-18", "PHYL25_2023-08-14", "PHYL25_2023-08-15", "PHYL25_2023-08-16", "PHYL25_2023-08-17", "PHYL25_2023-08-18", "PHYL28_2023-08-14", "PHYL28_2023-08-15", "PHYL34_2023-08-13", "PHYL34_2023-08-14",  "PHYL38_2023-08-12", "PHYL38_2023-08-14", "PHYL4_2023-08-14", "PHYL7_2023-08-16", "PHYL7_2023-08-17", "PHYL7_2023-08-18","PHYL9_2023-08-14",    "PHYL9_2023-08-18") == TRUE]
#27 need to calculate direction different, with the cave point and the first foraging location where stops for long
#"PHYL4_2023-08-15" direction is first to colony 3, calculate by hand?
#"PHYL9_2023-08-15"
#, "PHYL39_2023-08-14" removed
# "PHYL35_2023-08-14" removed
# PHYL4_2023-08-16 removed
#"PHYL16_2023-08-16", removed
#"PHYL9_2023-08-17", removed
#"PHYL11_2023-08-14_1" removed
#9_18 is doing three commutes

# for loop to divide each day of the animal in commuting outbound and commuting inbound using around 30 minutes as a threshold value
mylist3 <- list() 
for (i in 1:length(commu_wet_complete)){
  x <- which(commu_wet_complete[[i]]$diff>=1900)
  mylist3[[i]] <- split(commu_wet_complete[[i]], cumsum(1:nrow(commu_wet_complete[[i]])%in%x))
  indv <- unique(names(commu_wet_complete[i]))
  names(mylist3[[i]]) <- paste(indv, seq_along(mylist3[[i]]), sep = "_")
}

# convert nested list to a data frame
data_wet <- find_df(mylist3)

# add names to list of the commuting outbound or inbound to the list
data_wet <- Map(f, data_wet, names(data_wet))

# convert list to data frame
data.commu.wet <- as.data.frame(do.call(rbind, data_wet))

# select only outbound commuting
data.commu.wet.out <- data.commu.wet[grep("_1\\b", data.commu.wet$name),]#exact match to avoid selecting the #inbound commuting tracks
data.commu.wet.in <- data.commu.wet[grep("_2\\b", data.commu.wet$name),]

# plot lat and long
ggplot(aes(x=location_long, y=location_lat, color=tag_local_identifier), data=data.commu.wet.out)+geom_point()+ggtitle(data.commu.wet.out$ID_batday)#they are OK
```

## Combining commuting data frames and calculate straightness index

```{r, ugly-table, results=FALSE}
# merge data frames
all_comm_out <- rbind(data.commu.out,data.commu.march.out,data.commu.wet.out)

# straightness index
dist_traj<- lapply(split(all_comm_out,all_comm_out$ID), function(x){
 pd <-  pointDistance(head(x[,c("location_long","location_lat")],1),tail(x[,c("location_long","location_lat")], 1), lonlat = T)
 y <-  sum(x$step, na.rm=TRUE)
 st_idx <- c(id_day=unique(x$ID), id=unique(x$tag_local_identifier), cave=unique(x$year_cave), dist= pd, sum_step=y, straightness=pd/y)
 return(st_idx)
}
)

dist_traj_df <- as.data.frame(do.call(rbind, dist_traj))

# summary by ID
mean_straight_id <- dist_traj_df %>%
  group_by(id, cave)%>%
  dplyr::summarise(mean_st=mean(as.numeric(straightness),na.rm=TRUE),sd_st=sd(as.numeric(straightness), na.rm=TRUE))

# summary by cave
mean_straight_cave<- dist_traj_df %>%
  group_by(cave)%>%
   dplyr::summarise(mean_st=mean(as.numeric(straightness), na.rm=TRUE),sd_st=sd(as.numeric(straightness), na.rm=TRUE))

#count the number of nights per colony in the different years
a <- as.data.frame(table(all_comm_out$year_cave,all_comm_out$ID_batday))
b <- count(a$Var1[which(a$Freq>0)])
```

In general all bats in all periods and caves forage straight distances
